Refactor decision vs merge code
